django/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
Authlib==1.0.0
base58==2.1.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.4
cryptography==36.0.2
defusedxml==0.7.1
Django==4.0.2
django-neomodel==0.0.7
djangorestframework==3.11.2
frozenlist==1.3.0
gunicorn==20.1.0
idna==3.3
Jinja2==3.1.1
lxml==4.8.0
MarkupSafe==2.1.1
multidict==6.0.2
neo4j-driver==4.3.1
neobolt==1.7.17
neomodel==4.0.4
oauthlib==3.2.0
Pillow==9.1.0
prettytable==3.3.0
psycopg2-binary==2.9.3
pycparser==2.21
pycryptodome==3.14.1
pyecharts==1.9.1
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.1
requests==2.27.1
requests-oauthlib==1.3.1
Shapely==1.7.1
simplejson==3.17.6
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
tqdm==4.64.0
tzdata==2021.5
urllib3==1.26.9
wcwidth==0.2.5
whitenoise==3.3.1
yarl==1.7.2
